+++ src/test/resources/testbash/tests/fail.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 set -x
 
-echo ">>> This test should always fail"
+echo ">>> This test should always fail but that is to be expected"
 ./testbash "missingresource" --real_number abc --whole_number abc -s "a string with spaces" --truth \
   --output ./output.txt --log ./log.txt \
   --optional foo --optional_with_default bar